Break menus into their own files.
