+++ packages/mainmenu/src/index.ts
@@ -2,4 +2,10 @@
 // Distributed under the terms of the Modified BSD License.
 
 export * from './mainmenu';
-export * from './menus';
+export * from './labmenu';
+export * from './edit';
+export * from './file';
+export * from './help';
+export * from './kernel';
+export * from './run';
+export * from './view';
