+++ packages/mainmenu/src/run.ts
@@ -1,0 +1,31 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+import {
+  Menu
+} from '@phosphor/widgets';
+
+import {
+  IJupyterLabMenu, JupyterLabMenu
+} from './labmenu';
+
+/**
+ * An interface for a Run menu.
+ */
+export
+interface IRunMenu extends IJupyterLabMenu {
+}
+
+/**
+ * An extensible Run menu for the application.
+ */
+export
+class RunMenu extends JupyterLabMenu implements IRunMenu {
+  /**
+   * Construct the run menu.
+   */
+  constructor(options: Menu.IOptions) {
+    super(options);
+    this.title.label = 'Run';
+  }
+}
